Refactor NULS Address class
