+++ swift/Sources/Addresses/CoinType+Address.swift
@@ -65,6 +65,8 @@ public extension CoinType {
             }
         case .iost:
             return IOSTAccount(string: string)
+        case .nuls:
+            return NULSAddress(string: string)
         }
         return .none
     }
